+++ sandpack-client/src/clients/runtime/index.ts
@@ -98,7 +98,7 @@ export class SandpackRuntime extends SandpackClient {
 
       this.iframe.setAttribute(
         "allow",
-        "accelerometer; camera; encrypted-media; geolocation; gyroscope; hid; microphone; midi; clipboard-write;"
+        "accelerometer; camera; encrypted-media; geolocation; gyroscope; hid; microphone; midi; clipboard-read; clipboard-write;"
       );
     }